requirements.txt
flask 
dash
psycopg2-binary
pandas
flask-bcrypt
dash-bootstrap-components
python-dotenv
setuptools
flask_login
flask_sqlalchemy
